refactor: relocate group of functions to nuxt.ts for clarity
